upgrade ts-codegen and use bundles
