+++ cli/types/contracts/Govec.types.ts
@@ -1,0 +1,158 @@
+/**
+* This file was automatically generated by @cosmwasm/ts-codegen@0.10.0.
+* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
+*/
+
+export interface AllAccountsResponse {
+  accounts: string[];
+  [k: string]: unknown;
+}
+export type Uint128 = string;
+export interface BalanceResponse {
+  balance: Uint128;
+  [k: string]: unknown;
+}
+export type Binary = string;
+export type DaoResponse = string;
+export interface DownloadLogoResponse {
+  data: Binary;
+  mime_type: string;
+  [k: string]: unknown;
+}
+export type ExecuteMsg = {
+  transfer: {
+    amount: Uint128;
+    recipient: string;
+    [k: string]: unknown;
+  };
+} | {
+  burn: {
+    [k: string]: unknown;
+  };
+} | {
+  send: {
+    amount: Uint128;
+    contract: string;
+    msg: Binary;
+    [k: string]: unknown;
+  };
+} | {
+  mint: {
+    new_wallet: string;
+    [k: string]: unknown;
+  };
+} | {
+  update_staking_addr: {
+    new_addr: string;
+    [k: string]: unknown;
+  };
+} | {
+  update_mint_data: {
+    new_mint?: MinterData | null;
+    [k: string]: unknown;
+  };
+} | {
+  update_dao_addr: {
+    new_addr: string;
+    [k: string]: unknown;
+  };
+} | {
+  update_marketing: {
+    description?: string | null;
+    marketing?: string | null;
+    project?: string | null;
+    [k: string]: unknown;
+  };
+} | {
+  upload_logo: Logo;
+};
+export type Logo = {
+  url: string;
+} | {
+  embedded: EmbeddedLogo;
+};
+export type EmbeddedLogo = {
+  svg: Binary;
+} | {
+  png: Binary;
+};
+export interface MinterData {
+  cap?: Uint128 | null;
+  minter: string;
+  [k: string]: unknown;
+}
+export type LogoInfo = "embedded" | {
+  url: string;
+};
+export type Addr = string;
+export interface InstantiateMsg {
+  initial_balances: Cw20Coin[];
+  marketing?: MarketingInfoResponse | null;
+  minter?: MinterData | null;
+  name: string;
+  staking_addr?: string | null;
+  symbol: string;
+  [k: string]: unknown;
+}
+export interface Cw20Coin {
+  address: string;
+  amount: Uint128;
+  [k: string]: unknown;
+}
+export interface MarketingInfoResponse {
+  description?: string | null;
+  logo?: LogoInfo | null;
+  marketing?: Addr | null;
+  project?: string | null;
+  [k: string]: unknown;
+}
+export interface MinterResponse {
+  cap?: Uint128 | null;
+  minter: string;
+  [k: string]: unknown;
+}
+export type QueryMsg = {
+  balance: {
+    address: string;
+    [k: string]: unknown;
+  };
+} | {
+  token_info: {
+    [k: string]: unknown;
+  };
+} | {
+  minter: {
+    [k: string]: unknown;
+  };
+} | {
+  staking: {
+    [k: string]: unknown;
+  };
+} | {
+  dao: {
+    [k: string]: unknown;
+  };
+} | {
+  all_accounts: {
+    limit?: number | null;
+    start_after?: string | null;
+    [k: string]: unknown;
+  };
+} | {
+  marketing_info: {
+    [k: string]: unknown;
+  };
+} | {
+  download_logo: {
+    [k: string]: unknown;
+  };
+};
+export type StakingResponse = string;
+export interface TokenInfoResponse {
+  decimals: number;
+  name: string;
+  symbol: string;
+  total_supply: Uint128;
+  [k: string]: unknown;
+}
