+++ cli/types/contracts/index.ts
@@ -1,0 +1,23 @@
+/**
+* This file was automatically generated by @cosmwasm/ts-codegen@0.10.0.
+* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
+*/
+
+import * as _0 from "./Proxy.types";
+import * as _1 from "./Proxy.client";
+import * as _2 from "./Factory.types";
+import * as _3 from "./Factory.client";
+import * as _4 from "./Govec.types";
+import * as _5 from "./Govec.client";
+export namespace contracts {
+  export const Proxy = { ..._0,
+    ..._1
+  };
+  export const Factory = { ..._2,
+    ..._3
+  };
+  export const Govec = { ..._4,
+    ..._5
+  };
+}
