+++ cli/types/package.json
@@ -12,9 +12,7 @@
         "publish:dev": "tsc && cp package.json build/package.json && cd build && npm publish --dry-run",
         "publish:prod": "tsc && cp package.json build/package.json && cd build && npm publish",
         "generate": "npm run types:proxy && npm run types:factory && npm run types:govec && npm version patch",
-        "types:proxy": "cosmwasm-ts-codegen generate --schema ../../contracts/proxy/schema --out ./contracts --name Proxy",
-        "types:factory": "cosmwasm-ts-codegen generate --schema ../../contracts/factory/schema --out ./contracts --name Factory",
-        "types:govec": "cosmwasm-ts-codegen generate --schema ../../contracts/govec/schema --out ./contracts --name Govec"
+        "types": "node ./scripts/gen-contracts.js"
     },
     "license": "Apache-2.0",
     "homepage": "https://www.nymlab.it",
@@ -29,7 +27,7 @@
     "dependencies": {
         "@cosmjs/amino": "^0.28.7",
         "@cosmjs/cosmwasm-stargate": "^0.28.7",
-        "@cosmwasm/ts-codegen": "^0.7.4",
+        "@cosmwasm/ts-codegen": "^0.10.0",
         "typescript": "^4.7.3"
     }
 }
